Add explicit root disk to Bastion host Machine
Instead of RAM booting the bastion host from the corresponding OS image,
we will be using a dedicated root disk which will be defaulted to 10Gi.
A new configuration parameter has been added to the Bastion host
configuration in called `VolumeClassName` which indicates which
VolumeClass to use when creating the Bastion host.